move activity-related components into activities subfolder
