+++ client/src/components/Today/Today.tsx
@@ -1,7 +1,7 @@
+import NewActivity from "@/components/activities/NewActivity/NewActivity";
 import Calendar from "@/components/Calendar/Calendar";
 import Habits from "@/components/habits/Habits/Habits";
 import Modal from "@/components/Modal";
-import NewActivity from "@/components/NewActivity/NewActivity";
 import NewHabit from "@/components/NewHabit/NewHabit";
 import AllDayActivities from "@/components/Today/AllDayActivities";
 import ChangeDayButton from "@/components/Today/ChangeDayButton";
