+++ client/src/components/activities/ActivityList/get-date-field.ts
