+++ client/src/components/activities/NewActivity/NewActivity.style.tsx
