+++ client/src/lib/router.tsx
@@ -10,8 +10,10 @@ import { createBrowserRouter } from "react-router-dom";
 const Home = lazy(() => import("@components/Home"));
 const NewNote = lazy(() => import("@components/NewNote/NewNote"));
 const Register = lazy(() => import("@components/Register/Register"));
-const NewActivity = lazy(() => import("@components/NewActivity/NewActivity"));
-const ActivityList = lazy(() => import("@components/ActivityList/ActivityList"));
+const NewActivity = lazy(() => import("@components/activities/NewActivity/NewActivity"));
+const ActivityList = lazy(
+	() => import("@/components/activities/ActivityList/ActivityList")
+);
 const Today = lazy(() => import("@components/Today/Today"));
 const Notes = lazy(() => import("@components/Notes/Notes"));
